Relocated ReplaceHeaderComponent component into a separate module
