+++ src/shell/components/RelationalFieldBase/FieldSelectorDialog/DialogHeader.tsx
@@ -5,11 +5,7 @@ import {
   IconButton,
   Button,
 } from "@mui/material";
-import {
-  CheckRounded,
-  CloseRounded,
-  AutorenewRounded,
-} from "@mui/icons-material";
+import { CheckRounded, CloseRounded } from "@mui/icons-material";
 
 type DialogHeaderProps = {
   selectedCount: number;
@@ -85,53 +81,6 @@ export const DialogHeader = ({
           startIcon={<CheckRounded />}
         >
           Done
-        </Button>
-        <IconButton size="small" onClick={onClose}>
-          <CloseRounded fontSize="small" />
-        </IconButton>
-      </Stack>
-    </DialogTitle>
-  );
-};
-
-type ReplaceDialogHeaderProps = Omit<
-  DialogHeaderProps,
-  "fieldLabel" | "multiselect" | "replace" | "onDeselectAll"
->;
-
-export const ReplaceDialogHeader = ({
-  selectedCount,
-  onClose,
-  onDone,
-  loading,
-}: ReplaceDialogHeaderProps) => {
-  return (
-    <DialogTitle
-      component="div"
-      sx={{
-        pt: 4,
-        pb: 2,
-        px: 4,
-        display: "flex",
-        justifyContent: "space-between",
-        alignItems: "center",
-      }}
-    >
-      <Typography variant="h3" fontWeight={700} data-cy="selected-count">
-        {!selectedCount || loading
-          ? "Select Replacement Item"
-          : "Replacement Item Selected"}
-      </Typography>
-      <Stack direction="row" gap={1}>
-        <Button
-          data-cy="done-selecting-item-button"
-          size="small"
-          variant="contained"
-          onClick={onDone}
-          startIcon={<AutorenewRounded />}
-          disabled={!selectedCount || loading}
-        >
-          Replace
         </Button>
         <IconButton size="small" onClick={onClose}>
           <CloseRounded fontSize="small" />
