+++ src/shell/components/RelationalFieldBase/FieldSelectorDialog/index.tsx
@@ -35,7 +35,8 @@ import { VersionCell } from "./VersionCell";
 import { ItemsLoading } from "./ItemsLoading";
 import { useGetUsersQuery } from "../../../services/accounts";
 import { NoSearchResults } from "../../NoSearchResults";
-import { DialogHeader, ReplaceDialogHeader } from "./DialogHeader";
+import { DialogHeader } from "./DialogHeader";
+import { ReplaceDialogHeader } from "./ReplaceDialogHeader";
 import { fetchItems } from "../../../store/content";
 import { AppState } from "../../../store/types";
 import { ContentItem } from "../../../services/types";
